+++ baselines/acktr/defaults.py
@@ -1,0 +1,5 @@
+def mujoco():
+    return dict(
+        nsteps=2500,
+        value_network='copy'
+    )
