+++ setup.cfg
@@ -7,9 +7,3 @@ exclude =
     baselines/ddpg,
     baselines/ppo1,
     baselines/bench,
-    baselines/acktr, 
-    
-    
-    
-    
-
